+++ R/test.R
@@ -84,3 +84,19 @@ MASS::addterm
 #' @keywords internal
 #' @family tests
 NULL
+
+#' Test case: widgets
+#'
+#' @name test-widgets
+#' @keywords internal
+#' @family tests
+#' @examples
+#' library(leaflet)
+#' leaflet() %>%
+#'   addTiles() %>%
+#'   addMarkers(
+#'     lng=174.768,
+#'     lat=-36.852,
+#'     popup="The birthplace of R"
+#'   )
+NULL